backend/requirements.txt
fastapi
uvicorn
websockets
psutil
py-cpuinfo